requirements.txt
playwright
fastapi
uvicorn[standard]
pydantic
pydoris
aiofiles
Jinja2
python-multipart
mysql-connector-python
httpx